requirements.txt
spade~=3.3.2
python-dotenv~=1.0.1
scikit-learn~=1.3.2
numpy~=1.24.4
pandas~=2.0.3
torch~=2.3.1
torchvision~=0.18.1
tqdm~=4.66.4
matplotlib~=3.7.5
numba~=0.58.1
PyYAML~=6.0.1
logging-test-case